Refactor hook cadence logic into separate module for improved testability
